pyproject.toml
[tool.poetry]
name = "redshift"
version = "0.1.0"
description = ""
authors = ["Andra Stroe <andra.stroe@cfa.harvard.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
astroquery = "^0.4"
astropy = "^4.0.1"
colorama = "^0.4.3"
numpy = "^1.18.4"
click = "^7.1.2"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
redshifts = 'redshifts.main:main'